chore: KuisNoticeApiClientRetryTest를 KuisNoticeApiClientTest로 이동
